- #110: moved techs data into techs/ folder.
