--- src/app/hanzhong/data/index.ts
@@ -1,1 +1,0 @@
-export { HANZHONG__WAR_TIERS__TECHS__BY_LEVEL, getTechById, getTechByName } from './hanzhong-techs';
