+++ src/app/hanzhong/techs/index.ts
@@ -5,3 +5,4 @@ export { isTechRequirementSatisfied } from './requirements';
 export type { HanzhongTechType } from './TechType';
 export type { HanzhongTechResourceType } from './TechResourceType';
 export type { HanzhongTechLevelType } from './TechLevelType';
+export { HANZHONG__WAR_TIERS__TECHS__BY_LEVEL, getTechById } from './techs--data';
