+++ src/app/hanzhong/techs/map-table-to-resources-and-bonuses.ts
