+++ src/app/hanzhong/utils/find-resources-needed-by-tech-id-and-level.ts
@@ -1,5 +1,4 @@
-import { getTechById } from '../data';
-import type { HanzhongTechResourceType } from '../techs';
+import { getTechById, type HanzhongTechResourceType } from '../techs';
 
 export const findResourcesNeededByTechIdAndLevel = (techId: string, level: number): HanzhongTechResourceType | -1 | null => {
   const tech = getTechById(techId);
