+++ src/app/hanzhong/war-tiers/war-tiers--data.ts
@@ -1,6 +1,6 @@
 import { sanitizeId } from '../../../utils';
 import { HANZHONG_WAR_TIER_IDS } from '../constants/items-ids';
-import { HANZHONG__WAR_TIERS__TECHS__BY_LEVEL } from '../data';
+import { HANZHONG__WAR_TIERS__TECHS__BY_LEVEL } from '../techs';
 
 import type { HanzhongWarTierType } from './WarTierType';
